final/requirements.txt
blinker==1.4
click==8.0.1
colorama==0.4.4
dateutils==0.6.12
Flask==2.0.1
Flask-Sitemap==0.3.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyfiglet==0.8.post1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
Werkzeug==2.0.1